python_service/csv_upload/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
cfenv==0.5.3
python-dotenv>=1.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.32.3
matplotlib==3.9.2
scikit-learn==1.5.1
numpy==2.1.1